Fix unit tests comparisions of Keyboard::Key
Linked: #62

+++ tests/integration/event/tests_Keyboard.cpp
@@ -156,7 +156,8 @@ TEST(Event, KeySequence)
 
     /// Initial state
     GTEST_ASSERT_FALSE(sequenceListener.isComplete());
-    GTEST_ASSERT_EQ(sequenceListener.getHeldKey(), event::Keyboard::Key::Unknown);
+    GTEST_ASSERT_EQ(
+        static_cast<size_t>(sequenceListener.getHeldKey()), static_cast<size_t>(event::Keyboard::Key::Unknown));
     GTEST_ASSERT_TRUE(sequenceListener.getValidatedKeys().empty());
 
     /// Call without force innaccessible
@@ -168,88 +169,100 @@ TEST(Event, KeySequence)
     /// First sequence key pressed
     event::EventsManager::handleEvent(registry, event::KeyPressedEvent(event::Keyboard::Key::A));
     GTEST_ASSERT_FALSE(sequenceListener.isComplete());
-    GTEST_ASSERT_EQ(sequenceListener.getHeldKey(), event::Keyboard::Key::A);
+    GTEST_ASSERT_EQ(static_cast<size_t>(sequenceListener.getHeldKey()), static_cast<size_t>(event::Keyboard::Key::A));
     GTEST_ASSERT_TRUE(sequenceListener.getValidatedKeys().empty());
     /// First sequence key released -> key is validated
     event::EventsManager::handleEvent(registry, event::KeyReleasedEvent(event::Keyboard::Key::A));
     GTEST_ASSERT_FALSE(sequenceListener.isComplete());
-    GTEST_ASSERT_EQ(sequenceListener.getHeldKey(), event::Keyboard::Key::Unknown);
-    GTEST_ASSERT_EQ(sequenceListener.getValidatedKeys()[0], event::Keyboard::Key::A);
+    GTEST_ASSERT_EQ(
+        static_cast<size_t>(sequenceListener.getHeldKey()), static_cast<size_t>(event::Keyboard::Key::Unknown));
+    GTEST_ASSERT_EQ(
+        static_cast<size_t>(sequenceListener.getValidatedKeys()[0]), static_cast<size_t>(event::Keyboard::Key::A));
 
     /// Following key without the handleEvent
     GTEST_ASSERT_FALSE(sequenceListener.update(event::KeyPressedEvent(event::Keyboard::Key::B)));
     GTEST_ASSERT_FALSE(sequenceListener.isComplete());
-    GTEST_ASSERT_EQ(sequenceListener.getHeldKey(), event::Keyboard::Key::B);
+    GTEST_ASSERT_EQ(static_cast<size_t>(sequenceListener.getHeldKey()), static_cast<size_t>(event::Keyboard::Key::B));
     GTEST_ASSERT_EQ(sequenceListener.getValidatedKeys().size(), 1);
 
     GTEST_ASSERT_FALSE(sequenceListener.update(event::KeyReleasedEvent(event::Keyboard::Key::B)));
     GTEST_ASSERT_FALSE(sequenceListener.isComplete());
-    GTEST_ASSERT_EQ(sequenceListener.getHeldKey(), event::Keyboard::Key::Unknown);
-    GTEST_ASSERT_EQ(sequenceListener.getValidatedKeys()[1], event::Keyboard::Key::B);
+    GTEST_ASSERT_EQ(
+        static_cast<size_t>(sequenceListener.getHeldKey()), static_cast<size_t>(event::Keyboard::Key::Unknown));
+    GTEST_ASSERT_EQ(
+        static_cast<size_t>(sequenceListener.getValidatedKeys()[1]), static_cast<size_t>(event::Keyboard::Key::B));
 
     /// Final key
     GTEST_ASSERT_FALSE(sequenceListener.update(event::KeyPressedEvent(event::Keyboard::Key::C)));
     GTEST_ASSERT_FALSE(sequenceListener.isComplete());
-    GTEST_ASSERT_EQ(sequenceListener.getHeldKey(), event::Keyboard::Key::C);
+    GTEST_ASSERT_EQ(static_cast<size_t>(sequenceListener.getHeldKey()), static_cast<size_t>(event::Keyboard::Key::C));
     GTEST_ASSERT_EQ(sequenceListener.getValidatedKeys().size(), 2);
     /// Sequence completed
     GTEST_ASSERT_TRUE(sequenceListener.update(event::KeyReleasedEvent(event::Keyboard::Key::C)));
     GTEST_ASSERT_TRUE(sequenceListener.isComplete());
-    GTEST_ASSERT_EQ(sequenceListener.getHeldKey(), event::Keyboard::Key::Unknown);
-    GTEST_ASSERT_EQ(sequenceListener.getValidatedKeys()[2], event::Keyboard::Key::C);
+    GTEST_ASSERT_EQ(
+        static_cast<size_t>(sequenceListener.getHeldKey()), static_cast<size_t>(event::Keyboard::Key::Unknown));
+    GTEST_ASSERT_EQ(
+        static_cast<size_t>(sequenceListener.getValidatedKeys()[2]), static_cast<size_t>(event::Keyboard::Key::C));
 
     /// Call is accessible and reset the sequence
     sequenceListener(registry, registry.getEntity(0));
     GTEST_ASSERT_EQ(sequenceCount, 2);
     GTEST_ASSERT_FALSE(sequenceListener.isComplete());
-    GTEST_ASSERT_EQ(sequenceListener.getHeldKey(), event::Keyboard::Key::Unknown);
+    GTEST_ASSERT_EQ(
+        static_cast<size_t>(sequenceListener.getHeldKey()), static_cast<size_t>(event::Keyboard::Key::Unknown));
     GTEST_ASSERT_TRUE(sequenceListener.getValidatedKeys().empty());
 
     /// Invalid first key pressed
     event::EventsManager::handleEvent(registry, event::KeyPressedEvent(event::Keyboard::Key::Z));
     GTEST_ASSERT_FALSE(sequenceListener.isComplete());
-    GTEST_ASSERT_EQ(sequenceListener.getHeldKey(), event::Keyboard::Key::Unknown);
+    GTEST_ASSERT_EQ(
+        static_cast<size_t>(sequenceListener.getHeldKey()), static_cast<size_t>(event::Keyboard::Key::Unknown));
     GTEST_ASSERT_TRUE(sequenceListener.getValidatedKeys().empty());
 
     /// First sequence key pressed
     event::EventsManager::handleEvent(registry, event::KeyPressedEvent(event::Keyboard::Key::A));
     GTEST_ASSERT_FALSE(sequenceListener.isComplete());
-    GTEST_ASSERT_EQ(sequenceListener.getHeldKey(), event::Keyboard::Key::A);
+    GTEST_ASSERT_EQ(static_cast<size_t>(sequenceListener.getHeldKey()), static_cast<size_t>(event::Keyboard::Key::A));
     GTEST_ASSERT_TRUE(sequenceListener.getValidatedKeys().empty());
 
     /// Key pressed while waiting release
     event::EventsManager::handleEvent(registry, event::KeyPressedEvent(event::Keyboard::Key::Z));
     GTEST_ASSERT_FALSE(sequenceListener.isComplete());
-    GTEST_ASSERT_EQ(sequenceListener.getHeldKey(), event::Keyboard::Key::Unknown);
+    GTEST_ASSERT_EQ(
+        static_cast<size_t>(sequenceListener.getHeldKey()), static_cast<size_t>(event::Keyboard::Key::Unknown));
     GTEST_ASSERT_TRUE(sequenceListener.getValidatedKeys().empty());
 
     /// First sequence key pressed and released
     event::EventsManager::handleEvent(registry, event::KeyPressedEvent(event::Keyboard::Key::A));
     event::EventsManager::handleEvent(registry, event::KeyReleasedEvent(event::Keyboard::Key::A));
     GTEST_ASSERT_FALSE(sequenceListener.isComplete());
-    GTEST_ASSERT_EQ(sequenceListener.getHeldKey(), event::Keyboard::Key::Unknown);
+    GTEST_ASSERT_EQ(
+        static_cast<size_t>(sequenceListener.getHeldKey()), static_cast<size_t>(event::Keyboard::Key::Unknown));
     GTEST_ASSERT_FALSE(sequenceListener.getValidatedKeys().empty());
 
     /// Invalid Key pressed in the sequence
     event::EventsManager::handleEvent(registry, event::KeyPressedEvent(event::Keyboard::Key::Z));
     GTEST_ASSERT_FALSE(sequenceListener.isComplete());
-    GTEST_ASSERT_EQ(sequenceListener.getHeldKey(), event::Keyboard::Key::Unknown);
+    GTEST_ASSERT_EQ(
+        static_cast<size_t>(sequenceListener.getHeldKey()), static_cast<size_t>(event::Keyboard::Key::Unknown));
     GTEST_ASSERT_TRUE(sequenceListener.getValidatedKeys().empty());
 
     /// Unhandled release: no held key -> no change
     event::EventsManager::handleEvent(registry, event::KeyReleasedEvent(event::Keyboard::Key::A));
     GTEST_ASSERT_FALSE(sequenceListener.isComplete());
-    GTEST_ASSERT_EQ(sequenceListener.getHeldKey(), event::Keyboard::Key::Unknown);
+    GTEST_ASSERT_EQ(
+        static_cast<size_t>(sequenceListener.getHeldKey()), static_cast<size_t>(event::Keyboard::Key::Unknown));
     GTEST_ASSERT_TRUE(sequenceListener.getValidatedKeys().empty());
 
     /// Unhandled release: key doesn't match the held key -> no change
     event::EventsManager::handleEvent(registry, event::KeyPressedEvent(event::Keyboard::Key::A));
     GTEST_ASSERT_FALSE(sequenceListener.isComplete());
-    GTEST_ASSERT_EQ(sequenceListener.getHeldKey(), event::Keyboard::Key::A);
+    GTEST_ASSERT_EQ(static_cast<size_t>(sequenceListener.getHeldKey()), static_cast<size_t>(event::Keyboard::Key::A));
     GTEST_ASSERT_TRUE(sequenceListener.getValidatedKeys().empty());
     event::EventsManager::handleEvent(registry, event::KeyReleasedEvent(event::Keyboard::Key::Z));
     GTEST_ASSERT_FALSE(sequenceListener.isComplete());
-    GTEST_ASSERT_EQ(sequenceListener.getHeldKey(), event::Keyboard::Key::A);
+    GTEST_ASSERT_EQ(static_cast<size_t>(sequenceListener.getHeldKey()), static_cast<size_t>(event::Keyboard::Key::A));
     GTEST_ASSERT_TRUE(sequenceListener.getValidatedKeys().empty());
 }
